Create http endpoint for resource entity (#7)
* ran go mod tidy

* defined the api definition for resources for get, list and update

* Added create and delete api definition in the openapi configuration

* Ran script to generated the http server code

* Implemented http server and implemented list resources

* Implemented create resouce http api

* Implemented deleteById http endpoint for resource

* Implemented getById for resource http endpoint

* Implemented resource update by id

* Fixed list resource adapter for http

* Implemented pagination in list resource api

* added edge case checks

---------

+++ internal/service/resource.go
@@ -40,14 +40,44 @@ func (r *Resource) getByName(resourceName string) (model.Resource, error) {
 }
 
 func (r *Resource) GetById(resourceId string) (model.Resource, error) {
-	return r.repository.GetById(resourceId)
+	resource, err := r.repository.GetById(resourceId)
+	if err != nil {
+		return model.Resource{}, err
+	}
+	if resource.Id == "" {
+		return model.Resource{}, util.ApiError{
+			HTTPErrorCode: 404,
+			GRPCErrorCode: codes.NotFound,
+			ErrorMessage:  "resource not found",
+		}
+	}
+	return resource, nil
 }
 
-func (r *Resource) List() ([]model.Resource, error) {
-	return r.repository.List()
+func (r *Resource) List(pageNumber, pageSize int) ([]model.Resource, error) {
+	offset := 0
+	if pageNumber > 1 {
+		offset = (pageNumber - 1) * pageSize
+	}
+	limit := 5
+	if pageSize > 0 {
+		limit = pageSize
+	}
+	return r.repository.List(limit, offset)
 }
 
 func (r *Resource) DeleteById(resourceId string) (string, error) {
+	existingResource, err := r.repository.GetById(resourceId)
+	if err != nil {
+		return "", err
+	}
+	if existingResource.Id == "" {
+		return "", util.ApiError{
+			HTTPErrorCode: 404,
+			GRPCErrorCode: codes.NotFound,
+			ErrorMessage:  "resource not found",
+		}
+	}
 	return r.repository.DeleteById(resourceId)
 }
